Fork register.html template and make necessary change
This is the only way I found to change things on the register page
that needed to be changed. We'll improve this later when we figure
out how exactly a template can be overridden without rewriting the
whole template inside the theme. This definitely needs revisiting
for a better solution.